+++ CHANGELOG.md
@@ -5,8 +5,7 @@ All notable changes to this project will be documented in this file.
 ## v2.2.3
 
 - ⚡️ Media streaming: Added 'custom' field for FFmpeg hardware acceleration transcoding settings
-- 🦺 Offline: Fixed some episode images not being downloaded
-  - If you encounter this issue, you will have to remove and re-add the media
+- 🦺 Offline: Fixed episode images
 - 🦺 Debrid streaming: Fixed MPV command
 - 🦺 Debrid streaming: Fixed stream cancellation
 - 🏗️ Moved filename parser to separate package